+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - improved handling of 'unknown' finish reason of some models when using Agent
 - added file viewer modal for Project Files
 - using no-verify during the rebase to avoid failure
+- added merge conflict indicators for updated files
 
 ## [0.61.1]